Release

+++ api/src/main/kotlin/handler/DeviceHandler.kt
@@ -1,18 +1,15 @@
 package com.wafflestudio.snutt.handler
 
 import com.wafflestudio.snutt.common.dto.OkResponse
+import com.wafflestudio.snutt.common.exception.InvalidPathParameterException
 import com.wafflestudio.snutt.middleware.SnuttRestApiDefaultMiddleware
 import com.wafflestudio.snutt.notification.service.DeviceService
-import com.wafflestudio.snutt.users.service.UserNicknameService
-import com.wafflestudio.snutt.users.service.UserService
 import org.springframework.stereotype.Component
 import org.springframework.web.reactive.function.server.ServerRequest
 
 @Component
 class DeviceHandler(
     private val deviceService: DeviceService,
-    private val userService: UserService,
-    private val userNicknameService: UserNicknameService,
     snuttRestApiDefaultMiddleware: SnuttRestApiDefaultMiddleware,
 ) : ServiceHandler(snuttRestApiDefaultMiddleware) {
     suspend fun addRegistrationId(req: ServerRequest) =
@@ -31,6 +28,7 @@ class DeviceHandler(
             val userId = req.userId
 
             val registrationId = req.pathVariable("id")
+            if (registrationId.isBlank()) throw InvalidPathParameterException("id")
             deviceService.removeRegistrationId(userId, registrationId)
 
             OkResponse()
